+++ src/brace_expand.h
@@ -1,0 +1,6 @@
+#ifndef BRACE_EXPAND_H
+#define BRACE_EXPAND_H
+
+char **expand_braces(const char *word, int *count_out);
+
+#endif /* BRACE_EXPAND_H */
